// cc/src/app/models/studentParam.ts
export interface StudentParam {
    _id: string;
    username: string;
    password: string;
    index: string;
    firstName: string;
    lastName: string;
    cuin: string;
    email: string;
    way: string;
    status: string;
    type: string;
    yearOfStart: Number;
    yearOfStudy: Number;
}